addHp: cast GetCID to unsigned short at each masked compare
Match the target's clrlwi+andi shape by casting GetCID() to unsigned
short before the & 0x6D tests. addHp 83.6 -> 85.2%.

+++ src/charaobj.cpp
@@ -2430,7 +2430,7 @@ int CGCharaObj::calcSta(int staIndex, int amount, CGObject* source)
  */
 void CGCharaObj::addHp(int delta, CGPrgObj* sourceObj)
 {
-	if ((GetCID() & 0x6D) == 0x6D &&
+	if ((static_cast<unsigned short>(GetCID()) & 0x6D) == 0x6D &&
 	    (MiniGamePcs.m_flags & 4) != 0) {
 		return;
 	}
@@ -2446,7 +2446,7 @@ void CGCharaObj::addHp(int delta, CGPrgObj* sourceObj)
 		}
 		*reinterpret_cast<unsigned short*>(reinterpret_cast<unsigned char*>(m_scriptHandle) + 0x1C) = static_cast<unsigned short>(next);
 	} else if (delta < 0 && hpValue != 0) {
-		if ((GetCID() & 0x6D) == 0x6D && (CFlatGameFlags() & CFlatGameFlag_Bit0) != 0 &&
+		if ((static_cast<unsigned short>(GetCID()) & 0x6D) == 0x6D && (CFlatGameFlags() & CFlatGameFlag_Bit0) != 0 &&
 		    static_cast<int>(hpValue + delta) <= 0) {
 			delta = -(static_cast<int>(hpValue) - 1);
 		}
@@ -2471,7 +2471,7 @@ void CGCharaObj::addHp(int delta, CGPrgObj* sourceObj)
 		*reinterpret_cast<unsigned short*>(reinterpret_cast<unsigned char*>(m_scriptHandle) + 0x1C) = static_cast<unsigned short>(next);
 		m_worldParam = kOneF32;
 
-		if ((GetCID() & 0x6D) == 0x6D) {
+		if ((static_cast<unsigned short>(GetCID()) & 0x6D) == 0x6D) {
 			CGPartyObj* party = static_cast<CGPartyObj*>(this);
 			if (static_cast<signed char>(party->m_partyData.partyFlags) < 0) {
 				int stackArgs[2];
